+++ src/main/java/br/ufpb/dcx/apps4society/educandoapi/repositories/ContextRepository.java
@@ -7,6 +7,6 @@ import br.ufpb.dcx.apps4society.educandoapi.domain.Context;
 
 
 @Repository
-public interface ContextRepository extends JpaRepository<Context, Integer> {
+public interface ContextRepository extends JpaRepository<Context, Long> {
 	
 }
